+++ src/py/flwr/common/typing.py
@@ -21,7 +21,13 @@ from typing import Dict, List, Optional, Union
 import numpy as np
 
 Weights = List[np.ndarray]
+
+# The following union type contains Python types corresponding to ProtoBuf types that
+# ProtoBuf considers to be "Scalar Value Types", even though some of them arguably do
+# not conform to other definitions of what a scalar is. Source:
+# https://developers.google.com/protocol-buffers/docs/overview#scalar
 Scalar = Union[bool, bytes, float, int, str]
+
 Metrics = Dict[str, Scalar]
 
 
@@ -45,8 +51,7 @@ class FitIns:
     """Fit instructions for a client."""
 
     parameters: Parameters
-    config: Dict[str, str]
-    # metrics: Metrics
+    config: Dict[str, Scalar]
 
 
 @dataclass
@@ -55,9 +60,9 @@ class FitRes:
 
     parameters: Parameters
     num_examples: int
-    num_examples_ceil: int  # Deprecated
-    fit_duration: float  # Deprecated
-    # metrics: Metrics
+    num_examples_ceil: Optional[int] = None  # Deprecated
+    fit_duration: Optional[float] = None  # Deprecated
+    metrics: Optional[Metrics] = None
 
 
 @dataclass
@@ -65,17 +70,16 @@ class EvaluateIns:
     """Evaluate instructions for a client."""
 
     parameters: Parameters
-    config: Dict[str, str]
-    # metrics: Metrics
+    config: Dict[str, Scalar]
 
 
 @dataclass
 class EvaluateRes:
     """Evaluate response from a client."""
 
-    num_examples: int
     loss: float
-    accuracy: float  # Deprecated
+    num_examples: int
+    accuracy: Optional[float] = None  # Deprecated
     metrics: Optional[Metrics] = None