+++ doc/source/changelog.rst
@@ -10,6 +10,62 @@ What's new?
     * (abstract) FedOpt
     * FedAdagrad
 
+Deprecations
+
+* Deprecate :code:`flwr.server.strategy.DefaultStrategy` (migrate to :code:`flwr.server.strategy.FedAvg`, which is equivalent)
+
+
+v0.14.0 (2021-02-18)
+--------------------
+
+What's new?
+
+* **Generalized** :code:`Client.fit` **and** :code:`Client.evaluate` **return values** (`#610 <https://github.com/adap/flower/pull/610>`_, `#572 <https://github.com/adap/flower/pull/572>`_, `#633 <https://github.com/adap/flower/pull/633>`_)
+
+  Clients can now return an additional dictionary mapping :code:`str` keys to values of the following types: :code:`bool`, :code:`bytes`, :code:`float`, :code:`int`, :code:`str`. This means one can return almost arbitrary values from :code:`fit`/:code:`evaluate` and make use of them on the server side!
+  
+  This improvement also allowed for more consistent return types between :code:`fit` and :code:`evaluate`: :code:`evaluate` should now return a tuple :code:`(float, int, dict)` representing the loss, number of examples, and a dictionary holding arbitrary problem-specific values like accuracy. 
+  
+  In case you wondered: this feature is compatible with existing projects, the additional dictionary return value is optional. New code should however migrate to the new return types to be compatible with upcoming Flower releases (:code:`fit`: :code:`List[np.ndarray], int, Dict[str, Scalar]`, :code:`evaluate`: :code:`float, int, Dict[str, Scalar]`). See the example below for details.
+
+  *Code example:* note the additional dictionary return values in both :code:`FlwrClient.fit` and :code:`FlwrClient.evaluate`: 
+
+  .. code-block:: python
+
+    class FlwrClient(fl.client.NumPyClient):
+        def fit(self, parameters, config):
+            net.set_parameters(parameters)
+            train_loss = train(net, trainloader)
+            return net.get_weights(), len(trainloader), {"train_loss": train_loss}
+
+        def evaluate(self, parameters, config):
+            net.set_parameters(parameters)
+            loss, accuracy, custom_metric = test(net, testloader)
+            return loss, len(testloader), {"accuracy": accuracy, "custom_metric": custom_metric}
+
+* **Generalized** :code:`config` **argument in** :code:`Client.fit` **and** :code:`Client.evaluate` (`#595 <https://github.com/adap/flower/pull/595>`_)
+
+  The :code:`config` argument used to be of type :code:`Dict[str, str]`, which means that dictionary values were expected to be strings. The new release generalizes this to enable values of the following types: :code:`bool`, :code:`bytes`, :code:`float`, :code:`int`, :code:`str`.
+  
+  This means one can now pass almost arbitrary values to :code:`fit`/:code:`evaluate` using the :code:`config` dictionary. Yay, no more :code:`str(epochs)` on the server-side and :code:`int(config["epochs"])` on the client side!
+
+  *Code example:* note that the :code:`config` dictionary now contains non-:code:`str` values in both :code:`Client.fit` and :code:`Client.evaluate`: 
+
+  .. code-block:: python
+  
+    class FlwrClient(fl.client.NumPyClient):
+        def fit(self, parameters, config):
+            net.set_parameters(parameters)
+            epochs: int = config["epochs"]
+            train_loss = train(net, trainloader, epochs)
+            return net.get_weights(), len(trainloader), {"train_loss": train_loss}
+
+        def evaluate(self, parameters, config):
+            net.set_parameters(parameters)
+            batch_size: int = config["batch_size"]
+            loss, accuracy = test(net, testloader, batch_size)
+            return loss, len(testloader), {"accuracy": accuracy}
+
 
 v0.13.0 (2021-01-08)
 --------------------